Updated useContractLoader to use more generic implementation found in useExternalContractLoader
