+++ src/ContractLoader.ts
@@ -5,78 +5,50 @@ import { Web3Provider, JsonRpcSigner } from "@ethersproject/providers";
 import { useState, useEffect } from "react";
 
 /*
+  when you want to load an existing contract using just the provider, address, and ABI
+*/
+
+/*
   ~ What it does? ~
 
-  Loads your local contracts and gives options to read values from contracts
-                                              or write transactions into them
+  Enables you to load an existing mainnet DAI contract using the provider, address and abi
 
   ~ How can I use? ~
 
-  const readContracts = useContractLoader(localProvider) // or
-  const writeContracts = useContractLoader(userProvider)
+  const mainnetDAIContract = useExternalContractLoader(mainnetProvider, DAI_ADDRESS, DAI_ABI)
 
   ~ Features ~
 
-  - localProvider enables reading values from contracts
-  - userProvider enables writing transactions into contracts
-  - Example of keeping track of "purpose" variable by loading contracts into readContracts
-    and using ContractReader.js hook:
-    const purpose = useContractReader(readContracts,"YourContract", "purpose")
-  - Example of using setPurpose function from our contract and writing transactions by Transactor.js helper:
-    tx( writeContracts.YourContract.setPurpose(newPurpose) )
+  - Specify mainnetProvider
+  - Specify DAI_ADDRESS and DAI_ABI, you can load/write them using constants.js
 */
-
-const loadContract = (
-  contractName: string,
-  signer: Web3Provider | JsonRpcSigner,
-  contractsPath: string
-): Contract => {
-  const newContract = new Contract(
-    require(require.main?.path + contractsPath + `/${contractName}.address.js`),
-    require(require.main?.path + contractsPath + `/${contractName}.abi.js`),
-    signer,
-  );
-  // try {
-  //   newContract.bytecode = require(`../contracts/${contractName}.bytecode.js`);
-  // } catch (e) {
-  //   console.log(e);
-  // }
-  return newContract;
-};
-
 export default function useContractLoader(
-  providerOrSigner: Web3Provider,
-  contractsPath: string = "../contracts"
-): any {
-  const [contracts, setContracts] = useState();
+  provider: Web3Provider,
+  address: string,
+  ABI: any,
+) {
+  const [contract, setContract] = useState<Contract>();
   useEffect(() => {
-    async function loadContracts() {
+    async function loadContract() {
       try {
-        // we need to check to see if this providerOrSigner has a signer or not
-        let signer: Web3Provider | JsonRpcSigner;
-        let accounts;
-        if (providerOrSigner && typeof providerOrSigner.listAccounts === "function") {
-          accounts = await providerOrSigner.listAccounts();
-        }
-
+        // we need to check to see if this provider has a signer or not
+        let signer;
+        const accounts = await provider.listAccounts();
         if (accounts && accounts.length > 0) {
-          signer = providerOrSigner.getSigner();
+          signer = provider.getSigner();
         } else {
-          signer = providerOrSigner;
+          signer = provider;
         }
 
-        const contractList = require(require.main?.path + contractsPath + "/contracts.js");
+        const customContract = new Contract(address, ABI, signer);
+        // if(optionalBytecode) customContract.bytecode = optionalBytecode
 
-        const newContracts = contractList.reduce((accumulator: any, contractName: string) => {
-          accumulator[contractName] = loadContract(contractName, signer, contractsPath);
-          return accumulator;
-        }, {});
-        setContracts(newContracts);
+        setContract(customContract);
       } catch (e) {
-        console.log("ERROR LOADING CONTRACTS!!", e);
+        console.log("ERROR LOADING EXTERNAL CONTRACT AT "+address+" (check provider, address, and ABI)!!", e);
       }
     }
-    loadContracts();
-  }, [providerOrSigner]);
-  return contracts;
+    loadContract();
+  }, [provider, address, ABI]);
+  return contract;
 }
