+++ src/index.ts
@@ -11,6 +11,6 @@ export { default as useTokenBalance } from "./TokenBalance";
 export { default as useUserAddress } from "./UserAddress";
 export { default as useUserProvider } from "./UserProvider";
 export { default as useContractLoader } from "./ContractLoader";
-export { default as useExternalContractLoader } from "./ExternalContractLoader";
+export { default as useLoaclContractsLoader } from "./LocalContractsLoader";
 export { default as useLocalStorage } from "./LocalStorage";
 export { default as useOnBlock } from "./OnBlock";
